Move declarations of genericEventNames & userSpecificEvents out of DB.ts
