+++ packages/augur-state/src/Controller.ts
@@ -1,5 +1,98 @@
-import { DB } from './db/DB';
+import { DB, UserSpecificEvent } from './db/DB';
 import { Augur } from 'augur-api';
+
+// TODO Get these from GenericContractInterfaces (and do not include any that are unneeded)
+const genericEventNames: Array<string> = [
+  "DisputeCrowdsourcerCompleted",
+  "DisputeCrowdsourcerCreated",
+  "DisputeWindowCreated",
+  "MarketCreated",
+  "MarketFinalized",
+  "MarketMigrated",
+  "MarketParticipantsDisavowed",
+  "ReportingParticipantDisavowed",
+  "TimestampSet",
+  "TokensBurned",
+  "TokensMinted",
+  "UniverseCreated",
+  "UniverseForked",
+];
+
+// TODO Update numAdditionalTopics/userTopicIndexes once contract events are updated
+const userSpecificEvents: Array<UserSpecificEvent> = [
+  {
+    "name": "CompleteSetsPurchased",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 2,
+  },
+  {
+    "name": "CompleteSetsSold",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 2,
+  },
+  {
+    "name": "DisputeCrowdsourcerContribution",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 1,
+  },
+  {
+    "name": "DisputeCrowdsourcerRedeemed",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 1,
+  },
+  {
+    "name": "InitialReporterRedeemed",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 1,
+  },
+  {
+    "name": "InitialReportSubmitted",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 1,
+  },
+  {
+    "name": "InitialReporterTransferred", 
+    "numAdditionalTopics": 2,
+    "userTopicIndex": 2,
+  },
+  {
+    "name": "MarketMailboxTransferred",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 2,
+  },
+  {
+    "name": "MarketTransferred",
+    "numAdditionalTopics": 2,
+    "userTopicIndex": 1,
+  },
+  {
+    "name": "OrderCanceled",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 2,
+  },
+  {
+    "name": "OrderCreated",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 0,
+  },
+  {
+    "name": "OrderFilled",
+    "numAdditionalTopics": 2,
+    "userTopicIndex": 1,
+  },
+  {
+    "name": "TokensTransferred",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 2,
+  },
+  {
+    "name": "TradingProceedsClaimed",
+    "numAdditionalTopics": 3,
+    "userTopicIndex": 2,
+  },
+];
+
+
 
 export class Controller<TBigNumber> {
   private dbController: DB<TBigNumber>;
@@ -17,7 +110,7 @@ export class Controller<TBigNumber> {
 
   public async run(): Promise<void> {
     try {
-      this.dbController = await DB.createAndInitializeDB(this.networkId, this.defaultStartSyncBlockNumber, this.trackedUsers);
+      this.dbController = await DB.createAndInitializeDB(this.networkId, this.defaultStartSyncBlockNumber, this.trackedUsers, genericEventNames, userSpecificEvents);
       await this.dbController.sync(this.augur, 100000, 5);
 
       // TODO Move this function into separate test
